Update documentation for query collection options
Clarify the conditions for automatic syncing and refetching, discord context:
 
hazn:  Is this code absolutely mandatory in my `queryCollectionOptions` to refetch using a livequery? It didn't seem to work before:  
```js
refetchInterval: 5000, // Poll every 5 seconds to sync with server
staleTime: 0, // Data is immediately considered stale
```

Kyle Mathews:  The `staleTime` isn't necessary — but yeah, it won't poll for updates otherwise.

hazn:  Thanks! I'll create a quick PR mentioning this, as it took me a while to figure this out. I expected the default to be automatically refetching.

Kyle Mathews:  Ah yeah sure — we copy the Query defaults for this.